lots of bugfixing and optimizing
- Split util into util and cache functions
- Optimized update and extract functions
- Added "dry run" debugging function, json dumps
- Removed unused and obsolete util functions

+++ src/extract.py
@@ -1,34 +1,27 @@
 import shutil
-import pandas as pd
 import os.path
-from pathlib import Path
-import config, util, update, system
+import config, cache, system, util
 
-def extract_projects(listprojects):
-    
-    cfg = config.require_config()
-    ud = update.require_update()
+def extract_projects(listprojects: list[str]):
+    cfg = config.get_global('config')
+    ud = config.get_global('update')
+    dryrun = config.get_global('dryrun')
 
     # ensure cache is up to date
-    ud.update_all()
+    ud.update_cache()
 
-    # check if we've generated our cache files
-    if not os.path.exists(cfg.cache / config.animdata_csv_path) or not os.path.exists(cfg.cache / config.animsetdata_csv_path):
-        print("Error: Could not update program.")
-        return None
-
-    animdata_csv = pd.read_csv(cfg.cache / config.animdata_csv_path)
-    animsetdata_csv = pd.read_csv(cfg.cache / config.animsetdata_csv_path)
+    animdata_df = ud.animdata_df
+    animsetdata_df = ud.animsetdata_df
 
     for project in listprojects:
         project = project.lower()
 
         # check if project exists
-        if util.is_in_cache(project) == False:
-            print(f"Warning: Project {project} not found. Skipping...")
-            continue
+        if cache.is_in_cache(project) == False:
+            print(f"Warning: Project {project} not found. Cancelling...")
+            return 0
 
-        unpacked = util.is_unpacked(project)
+        unpacked = cache.is_unpacked(project)
 
         # check if there are already extracted files in meshes
         if unpacked[0] == True or unpacked[1] == True or unpacked[2] == True:
@@ -54,10 +47,7 @@ def extract_projects(listprojects):
 
         if ud.creature_projects.__contains__(project):
             isCreature = True
-            animset_index = ud.animsetdata_csv.index[ud.animsetdata_csv['animset_name'] == project].to_list()[0]
-
-        print("Project index for {} is {}. Animset index is {}.".format(project, project_index, animset_index))
-
+            animset_index = animsetdata_df.index[animsetdata_df['animset_name'] == project].to_list()[0]
 
         ### ANIMATIONDATA ###
 
@@ -82,17 +72,19 @@ def extract_projects(listprojects):
         # open the animdata file
         with open(cfg.skyrim / config.animdata, "r", encoding="utf-8") as readable:
 
+            readline = readable.readline
+            strip = str.strip
+
             # write the animdata cache file
             try:
                 with open((animdata_temp_folder / (project + ".txt")), 'w', encoding="utf-8") as animdata_cache:
                     # read lines up to the project start
-                    for _ in range(int(animdata_csv.at[project_index, "project_start"])):
-                        readable.readline()
+                    util.fast_skip(readable, int(animdata_df.at[project_index, "project_start"]))
 
                     # skip line count
-                    readable.readline()
+                    util.fast_skip(readable, 1)
 
-                    lines_anims = int(animdata_csv.at[project_index, "lines_anims"])
+                    lines_anims = int(animdata_df.at[project_index, "lines_anims"])
 
                     n = 1
                     
@@ -100,8 +92,8 @@ def extract_projects(listprojects):
                         n = 2
 
                     for i in range(lines_anims - n):
-                        animdata_cache.write(readable.readline())
-                    animdata_cache.write(readable.readline().strip())
+                        animdata_cache.write(readline())
+                    animdata_cache.write(strip(readline()))
 
             except Exception as e:
                 print(f"Error extracting animdata for {project}: {e}")
@@ -119,16 +111,15 @@ def extract_projects(listprojects):
                     with open(boundanims_dir / ("anims_" + project + ".txt"), 'w', encoding="utf-8") as boundanims_cache:
 
                         # skip whitespace and linecount
-                        readable.readline()
-                        readable.readline()
+                        util.fast_skip(readable, 2)
 
-                        lines_boundanims = int(animdata_csv.at[project_index, "lines_boundanims"])
+                        lines_boundanims = int(animdata_df.at[project_index, "lines_boundanims"])
 
                         for i in range(lines_boundanims - 2):
 
-                            boundanims_cache.write(readable.readline())
+                            boundanims_cache.write(readline())
 
-                        boundanims_cache.write(readable.readline().strip())
+                        boundanims_cache.write(strip(readline()))
 
 
                 except Exception as e:
@@ -143,15 +134,17 @@ def extract_projects(listprojects):
             # open the animdata file
             with open(cfg.skyrim / config.animsetdata, "r", encoding="utf-8") as readable:
 
+                readline = readable.readline
+                strip = str.strip
+
                 # write the animsetdata cache file
                 try:
                     # skip to the project start
-                    for _ in range(int(ud.animsetdata_csv.at[animset_index, "animset_start"])):
-                        readable.readline()
+                    util.fast_skip(readable, int(animsetdata_df.at[animset_index, "animset_start"]))  # skip line count
 
                     # get the expected number of animation sets
-                    animset_count = int(readable.readline().strip())
-                    expected_animset_count = int(ud.animsetdata_csv.at[animset_index, "count_animsets"])
+                    animset_count = int(strip(readline()))
+                    expected_animset_count = int(animsetdata_df.at[animset_index, "count_animsets"])
 
                     # check to make sure the cache data matches
                     if animset_count != expected_animset_count:
@@ -172,7 +165,7 @@ def extract_projects(listprojects):
 
                         # write the file names to the animsetlist txt
                         for i in range(animset_count):
-                            animset_list.append(readable.readline().strip().lower())
+                            animset_list.append(strip(readline()).lower())
 
                         for item in animset_list:
                             animsetlist.write(item + "\n")
@@ -181,30 +174,29 @@ def extract_projects(listprojects):
                         with open(project_dir / animset_list[i], 'w', encoding="utf-8") as writable:
 
                             # write the first V3
-                            writable.write(readable.readline())
+                            writable.write(readline())
                         
                             # write notes A, B, and C
-                            notes_A = readable.readline()
-                            notes_A_int = int(notes_A.strip())
-
+                            notes_A = readline()
+                            notes_A_int = int(strip(notes_A))
                             writable.write(notes_A)
 
                             if notes_A_int != 0:
                                 for j in range(notes_A_int):
-                                    writable.write(readable.readline())
+                                    writable.write(readline())
 
-                            notes_B = readable.readline()
-                            notes_B_int = int(notes_B.strip())
+                            notes_B = readline()
+                            notes_B_int = int(strip(notes_B))
 
                             writable.write(notes_B)
                             
                             if notes_B_int != 0:
                                 for j in range(notes_B_int):
                                     for k in range(3):
-                                        writable.write(readable.readline())
+                                        writable.write(readline())
 
-                            notes_C = readable.readline()
-                            notes_C_int = int(notes_C.strip())
+                            notes_C = readline()
+                            notes_C_int = int(strip(notes_C))
 
                             writable.write(notes_C)
 
@@ -212,30 +204,30 @@ def extract_projects(listprojects):
                                 for j in range(notes_C_int):
                                     # skip two lines
                                     for k in range(2):
-                                        writable.write(readable.readline())
+                                        writable.write(readline())
 
-                                    n = readable.readline()
+                                    n = readline()
                                     writable.write(n)
 
                                     # skip 
-                                    for k in range(int(n.strip())):
-                                        writable.write(readable.readline())
+                                    for k in range(int(strip(n))):
+                                        writable.write(readline())
 
-                            file_count = readable.readline()
-                            file_count_int = int(file_count.strip())
+                            file_count = readline()
+                            file_count_int = int(strip(file_count))
 
                             writable.write(file_count)
 
                             for j in range(file_count_int):
                                 # skip two lines
                                 for k in range(2):
-                                    writable.write(readable.readline())
+                                    writable.write(readline())
 
                                 # if we are on the last line, strip whitespace.
                                 if i == animset_count - 1 and j == file_count_int - 1:
-                                    writable.write(readable.readline().strip())
+                                    writable.write(strip(readline()))
                                 
-                                else: writable.write(readable.readline())
+                                else: writable.write(readline())
                       
                 except Exception as e:
                     print(f"Error while extracting animset data for {project}: {e}")
@@ -244,8 +236,12 @@ def extract_projects(listprojects):
                 
         print(f"Successfully extracted {project}.")
 
-    # move temp files to their final destination
-    shutil.copytree(cfg.cache / "temp", cfg.skyrim, dirs_exist_ok=True)
+    if not dryrun:
+        # move temp files to their final destination
+        shutil.copytree(cfg.cache / "temp", cfg.skyrim, dirs_exist_ok=True)
+    else:
+        print("Dry run complete. No changes were made.")
+        util.pause_wait_for_input()
 
     # remove temp folders if they still exist
     system.clean_temp()
@@ -253,16 +249,16 @@ def extract_projects(listprojects):
     return 0
 
 
-def extract_all(and_i_mean_all_of_them = False):
-    cfg = config.require_config()
-    ud = update.require_update()
+def extract_all(and_i_mean_all_of_them=False):
+    cfg = config.get_global('config')
+    ud = config.get_global('update')
 
     vanilla_projects_count = len(ud.vanilla_projects)
-    animdata_csv = pd.read_csv(cfg.cache / config.animdata_csv_path)
-    animsetdata_csv = pd.read_csv(cfg.cache / config.animsetdata_csv_path)
+    animdata_df = ud.animdata_df
+    animsetdata_df = ud.animsetdata_df
 
     # get count for total projects
-    total_projects = animdata_csv.shape[0]                  
+    total_projects = animdata_df.shape[0]                  
 
     # make sure we have modded projects to extract
     if total_projects - vanilla_projects_count == 0 and not and_i_mean_all_of_them:
@@ -284,7 +280,7 @@ def extract_all(and_i_mean_all_of_them = False):
                         return 0
 
     # loop through all projects and check if they are vanilla or not
-    for project_name in (animdata_csv)['project_name']:
+    for project_name in (animdata_df)['project_name']:
         project_name = project_name.lower()
     
     to_extract = []
